requirements.txt
backports.ssl-match-hostname==3.5.0.1
bintrees==2.0.7
certifi==2017.4.17
chardet==3.0.4
gdax==1.0.6
PyYAML==3.12
requests==2.13.0
simplejson==3.11.1
six==1.10.0
urllib3==1.21.1
websocket-client==0.40.0
